experiment: consolidates tsup configs and build scripts
